Created GameMgr Class
The GameMgr Class is responsible for the rules of the game, as well as being able to transition to different scenes (to be added later).

This class is a static Class, so that its properties and methods are accessible to all other classes.

For now, it controls the world's gravity.